+++ src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/Vector128.cs
@@ -1838,7 +1838,7 @@ namespace System.Runtime.Intrinsics
             {
                 return IndexOf(vector.AsInt32(), -1);
             }
-            else if (typeof(T) == typeof(float))
+            else if (typeof(T) == typeof(double))
             {
                 return IndexOf(vector.AsInt64(), -1);
             }
@@ -2069,7 +2069,7 @@ namespace System.Runtime.Intrinsics
             {
                 return LastIndexOf(vector.AsInt32(), -1);
             }
-            else if (typeof(T) == typeof(float))
+            else if (typeof(T) == typeof(double))
             {
                 return LastIndexOf(vector.AsInt64(), -1);
             }
